config/input: store xkb layout index
This stores the xkb layout index in the input config for the device. The
stored value gets updated when the layout changes and reset when the
keymap is changed. The layout gets reapplied when configuring the
keyboard, as long as, either the layout did not change or the keyboard
did not previously have a layout set. Since input configs gets recreated
on reload, the layout also gets reset on reload.

This allows for the effective keyboard layout to be retained when the
keyboard is lost and re-discovered. This includes switching TTYs,
suspending and resuming, and manually unplugging and plugging the
keyboard back in. This commit also retains the keyboard layout when the
repeat info/delay are changed, but the keymap remains the same.